turned the audio services into classes
