+++ bigbluebutton-html5/imports/api/audio/client/bridge/base.js
@@ -1,0 +1,14 @@
+export default class BaseAudioService {
+  constructor() {
+  }
+
+  exitAudio () {
+  }
+
+  joinListenOnly() {
+  }
+
+  joinMicrophone() {
+  }
+
+}
