+++ bigbluebutton-html5/imports/ui/components/audio/service.js
@@ -3,8 +3,7 @@ import AudioModal from './audio-modal/component';
 import Meetings from '/imports/api/meetings';
 
 import { showModal } from '/imports/ui/components/app/service';
-import { joinListenOnly, joinMicrophone,
-  exitAudio } from '/imports/api/audio/client/main';
+import AudioAPI from '/imports/api/audio/client/bridge'
 
 const handleJoinAudio = () => {
   const handleJoinListenOnly = () => joinListenOnly();
@@ -14,6 +13,11 @@ const handleJoinAudio = () => {
 const getVoiceBridge = () => {
   return Meetings.findOne({}).voiceConf;
 } ;
+
+const audioAPI = new AudioAPI();
+let exitAudio = () => audioAPI.exitAudio();
+let joinListenOnly = () => audioAPI.joinListenOnly();
+let joinMicrophone = () => audioAPI.joinMicrophone();
 
 export {
   handleJoinAudio,
