+++ bigbluebutton-html5/imports/ui/components/deskshare/service.js
@@ -1,6 +1,8 @@
 import Deskshare from '/imports/api/deskshare';
-import {createVertoUserName, watchVertoVideo} from '/imports/api/audio/client/verto-bridge';
+import VertoAPI from '/imports/api/audio/client/bridge/verto';
 import Auth from '/imports/ui/services/auth';
+
+const vertoAPI = new VertoAPI();
 
 // when the meeting information has been updated check to see if it was
 // desksharing. If it has changed either trigger a call to receive video
@@ -28,12 +30,12 @@ function videoIsBroadcasting() {
 
 // if remote deskshare has been ended disconnect and hide the video stream
 function presenterDeskshareHasEnded() {
-  // exitVoiceCall();
+  // vertoAPI.exitVoiceCall();
 };
 
 // if remote deskshare has been started connect and display the video stream
 function presenterDeskshareHasStarted() {
-  vertoWatchVideo();
+  vertoAPI.vertoWatchVideo();
 };
 
 export {
